refactor(js): Move files to build inside a build folder.
